+++ tests/test_algorithms/test_rawDistanceEmbedder.py
@@ -9,7 +9,7 @@ class TestRawDistanceEmbedder(TestCase):
     def test__call__(self):
         # Arrange
         sut = RawDistanceEmbedder(max_pos=5)
-        expected = torch.tensor([[0, 1, 2, 3, 4, 5]])
+        expected = torch.tensor([[0, 1, 2, 3, 4, 5]]).float()
 
         # Act
         actual = sut()
